Объединяет файлы в один
